requirements.txt
python-dotenv==1.0.1
streamlit
langchain
langchain-community
langchain-huggingface
langchain-text-splitters==0.2.2
langchain-groq==0.1.9
transformers==4.43.3
sentence-transformers==3.0.1
faiss-cpu
unstructured[pdf]==0.15.0
unstructured==0.15.0
nltk